requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
bcrypt==4.2.0
cairocffi==1.7.1
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.3
Django==5.0.3
dnspython==2.3.0
et_xmlfile==2.0.0
Flask==2.2.2
Flask-Cors==4.0.1
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
minio==7.2.12
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
pycparser==2.22
pycryptodome==3.21.0
pygame==2.6.1
pymongo==4.3.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==0.21.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.1
watchdog==3.0.0
Werkzeug==2.2.2

django-qrcode~=0.3
qrcode~=8.2
PyJWT~=2.10.1minio
PyJWT
python-dotenv
